only unique recent searches

+++ scripts/lastfive.js
@@ -23,26 +23,33 @@ const mainContainer = document.getElementById('mainContainer')
 const addLocation = () => {
     const locVal = locationInput.value
     let key = 0
-    if (localStorage.length < 5) {
-        key = localStorage.length + 1
-        localStorage.setItem(key,locVal)
-    } else {
-        //removes earliest item
-        localStorage.removeItem(1)
-        //relabel keys
-        let keys = Object.keys(localStorage);
-        let tempObj = {}
-        for (let key of keys) {
-            tempObj[key-1] = localStorage.getItem(key)
-        }
-        localStorage.clear();
-        keys = Object.keys(tempObj);
-        let tempVal = ''
-        for (let key of keys) {
-            tempVal = tempObj[key];
-            localStorage.setItem(key,tempVal)
-        }
-        localStorage.setItem(5,locVal);
+    if (localStorage.length > 0) {
+        for (let i = 1; i< localStorage.length+1; i++) {
+            if (locVal === localStorage.getItem(i)) {
+                return
+            }
+        };
+        if (localStorage.length < 5) {
+            key = localStorage.length + 1
+            localStorage.setItem(key,locVal)
+        } else {
+            //removes earliest item
+            localStorage.removeItem(1)
+            //relabel keys
+            let keys = Object.keys(localStorage);
+            let tempObj = {}
+            for (let key of keys) {
+                tempObj[key-1] = localStorage.getItem(key)
+            }
+            localStorage.clear();
+            keys = Object.keys(tempObj);
+            let tempVal = ''
+            for (let key of keys) {
+                tempVal = tempObj[key];
+                localStorage.setItem(key,tempVal)
+            }
+            localStorage.setItem(5,locVal);
+        };    
     }
 }
